requirements.txt
redis==4.5.1
requests==2.28.2
requests_oauthlib==1.3.1
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
